UI for Buzz and cleanup of strings
Moves strings to strings.xml and adds the basic layout for buzz with placeholders for the text.

Todo:
Logic for timers for buzz
Connect FAB to firebase message and to animate changing color of the friends reminder.
Connect event names to firebase.
Shift between conencting to partner to the buzz layout.
Add images to the buzz layout and make it more playful
Gradually shift colors based on the timer.
Create animation when buzzed.
Notifation system when buzzed.