Setup new and existing event split off and use prerendered background.
Changes UI for new event
Adds existing event and button to choose between the two.
Adds a sample background as a place holder.

Todo:
Randomly generate background using the same object used for the buzz actvity expect send random values that slowly adjust over time to give an animated look.
Make status and nav bar transparent on this background.
Create logic for hiding and showing layouts based on which button is pressed.
Create logic for switching to the next view and conencting to firebase.
Create logic for checking ID in firebase.
Create logic for creating ID in firebase.