UI Changes and Activity Changes
Introduces Splash Screen where firebase will be connected from.
Once connected it will check for the ID of the device then if it exist send straight to Buzz with the correct settings, if it doesn't then it will sned to setup.

Redisgns the EditBoxes to make them have a background a and a ripple and removes the idea of multiple views in the same activity.