Fix Firebase
The problem before was where firebase was implemented.
Apple Plugin for GMS needs to be in the base which requires the google-services.json there as well but the implementation for the firebase components need to be in features

Creates DataManager calss to manage the data in an app wide way.