Reset Time Picker to current time every time it is opened.
Since TimeSelector is no longer staic the listener funtion is no longer called on init since the object isn't created on yet on constrution thus it is null meaning we can't fetch the mTimeText for the listener.
So to make behavior universal the dialog will always reset to the current time, before it was showing the current time even with the saved time in the box onCreate but after it was changed once it would then have the saved time every time it was opened after that, this would be confusing to the user as such it will just always reset to the current time.